Renamed clinic-ui to ClinicUI to follow pascal case for component everywhere.
