+++ tsconfig.app.json
@@ -5,7 +5,7 @@
       "src/*": ["./src/*"],
       "@pages": ["./src/pages/index"],
       "@schemas": ["./src/schemas/index"],
-      "@clinic-ui": ["./src/components/clinic-ui/index"],
+      "@clinic-ui": ["src/components/ClinicUI/index"],
       "@components/*": ["./src/components/*"],
       "@public/*": ["./public/*"],
       "@constants/*": ["./src/constants/*"]
